ConnectorWorker was moved to Conector class
